[MOBDW-218] Start refactoring initial screen
