+++ app/pages/Auth/CreateAccount.presenter/index.ts
@@ -1,0 +1,4 @@
+import { CreateAccountPresenter } from './CreateAccount.presenter';
+
+export default CreateAccountPresenter;
+export { CreateAccountPresenter };
