+++ app/pages/Auth/CreateAccount.view/index.ts
@@ -1,0 +1,4 @@
+import { CreateAccountView } from './CreateAccount.view';
+
+export default CreateAccountView;
+export { CreateAccountView };
