+++ app/pages/Auth/ImportAccount.presenter/index.ts
@@ -1,0 +1,4 @@
+import { ImportAccountPresenter } from './ImportAccount.presenter';
+
+export default ImportAccountPresenter;
+export { ImportAccountPresenter };
