+++ app/pages/Auth/ImportAccount.view/index.ts
@@ -1,0 +1,4 @@
+import { ImportAccountView } from './ImportAccount.view';
+
+export default ImportAccountView;
+export { ImportAccountView };
