+++ app/pages/Auth/UnlockWallet.presenter/index.ts
@@ -1,0 +1,4 @@
+import { UnlockWalletPresenter } from './UnlockWallet.presenter';
+
+export default UnlockWalletPresenter;
+export { UnlockWalletPresenter };
