+++ app/pages/Auth/UnlockWallet.view/index.ts
@@ -1,0 +1,4 @@
+import { UnlockWalletView } from './UnlockWallet.view';
+
+export default UnlockWalletView;
+export { UnlockWalletView };
