+++ app/pages/Auth/index.ts
@@ -1,0 +1,4 @@
+import { AuthPage } from './AuthPage.presenter';
+
+export default AuthPage;
+export { AuthPage };
