+++ app/pages/History/HistoryPage.presenter/HistoryPage.presenter.tsx
@@ -56,7 +56,6 @@ const HistoryPage: FC = () => {
     return [] as TransactionLog[];
   };
 
-  console.log(buildList());
   // CREATE VIEW
 
   if (transactionLogs === null) {
