--- app/views/auth/CreateAccountView/index.ts
@@ -1,3 +1,0 @@
-import CreateAccountView from './CreateAccountView';
-
-export default CreateAccountView;
