--- app/views/auth/ImportAccountView/index.ts
@@ -1,3 +1,0 @@
-import ImportAccountView from './ImportAccountView';
-
-export default ImportAccountView;
