--- app/views/auth/UnlockWalletView/index.ts
@@ -1,3 +1,0 @@
-import UnlockWalletView from './UnlockWalletView';
-
-export default UnlockWalletView;
