--- app/views/auth/index.ts
@@ -1,3 +1,0 @@
-export { default as CreateAccountView } from './CreateAccountView';
-export { default as ImportAccountView } from './ImportAccountView';
-export { default as UnlockWalletView } from './UnlockWalletView';
